+++ iOS/BrazeReactBridge/BrazeReactBridge/BrazeBannerManager.mm
@@ -5,8 +5,9 @@
 
 RCT_EXPORT_MODULE(BrazeBannerView)
 
-RCT_EXPORT_VIEW_PROPERTY(placementID, NSString *)
+RCT_EXPORT_VIEW_PROPERTY(placementId, NSString *)
 RCT_EXPORT_VIEW_PROPERTY(onHeightChanged, RCTBubblingEventBlock)
+RCT_EXPORT_VIEW_PROPERTY(onBannerDismiss, RCTBubblingEventBlock)
 
 + (BOOL)requiresMainQueueSetup {
   return YES;