+++ iOS/BrazeReactBridge/BrazeReactBridge/BrazeBannerUIHelper.h
@@ -10,8 +10,9 @@
 @interface BrazeBannerUIHelper : NSObject
 
 @property (nonatomic, copy) void (^_Nullable onHeightChanged)(double height);
+@property (nonatomic, copy) void (^_Nullable onDismiss)(NSString *placementId, NSString *stableKey, NSString *trackingId);
 
-- (void)insertPlacement:(NSString *_Nonnull)placementID
+- (void)insertPlacement:(NSString *_Nonnull)placementId
                intoView:(UIView *_Nonnull)hostView;
 - (void)resizeView:(UIView *_Nonnull)view withHeight:(NSNumber *_Nonnull)height;
 - (void)triggerHeightUpdate;