+++ src/specs/BrazeBannerViewNativeComponent.ts
@@ -2,14 +2,21 @@ import codegenNativeComponent from 'react-native/Libraries/Utilities/codegenNati
 import type { HostComponent, ViewProps } from 'react-native';
 import type { BubblingEventHandler, Double } from 'react-native/Libraries/Types/CodegenTypes';
 
-type BrazeBannerDimensionsEvent = {
+type BannerDimensionsEvent = {
   // Use `Double` from CodegenTypes because `number` types are not supported as a bubbling event type.
   height: Double;
 };
 
+type BannerDismissEvent = {
+  placementId: string;
+  stableKey: string;
+  trackingId: string;
+};
+
 export interface BrazeBannerViewProps extends ViewProps {
-  placementID: string;
-  onHeightChanged?: BubblingEventHandler<BrazeBannerDimensionsEvent> | null;
+  placementId: string;
+  onHeightChanged?: BubblingEventHandler<BannerDimensionsEvent> | null;
+  onBannerDismiss?: BubblingEventHandler<BannerDismissEvent> | null;
 }
 
 export default codegenNativeComponent<BrazeBannerViewProps>(